refactor: Reorganize project structure
